# src/main/resources/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Letras y Papeles API",
    "description": "Endpoints para la implementación del proyecto Letras y Papeles",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/order/{id}": {
      "get": {
        "tags": [
          " ** Endpoints Orden **"
        ],
        "summary": "Obtiene una orden por su ID",
        "description": "Este endpoint permite obtener una orden específica utilizando su ID",
        "operationId": "obtenerPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la orden",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Orden obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Orden no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          " ** Endpoints Orden **"
        ],
        "summary": "Actualiza una orden existente",
        "description": "Este endpoint permite actualizar los detalles de una orden existente",
        "operationId": "actualizarOrden",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la orden",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Datos de la orden a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Orden actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Orden no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/branch/{id}": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Obtiene una sucursal por su ID",
        "description": "Recupera los detalles de una sucursal específica utilizando su identificador único",
        "operationId": "obtenerPorId_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la sucursal",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Sucursal encontrada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "description": "Sucursal no encontrada con el ID proporcionado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "ID de sucursal inválido.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Actualiza una sucursal existente",
        "description": "Permite modificar completamente los datos de una sucursal existente identificándola por su ID",
        "operationId": "actualizarSucursal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la sucursal a actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Datos actualizados de la sucursal",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Branch"
              },
              "examples": {
                "Sucursal actualizada": {
                  "summary": "Ejemplo de actualización de sucursal",
                  "description": "Sucursal actualizada",
                  "value": {
                    "name": "Sucursal Test Actuliazada",
                    "address": "Av. Test 123",
                    "city": "Ciudad de Test"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sucursal actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "description": "Sucursal no encontrada con el ID proporcionado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "Datos de la sucursal inválidos o incompletos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Elimina una sucursal",
        "description": "Permite eliminar una sucursal específica del sistema utilizando su identificador único",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la sucursal a eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Sucursal eliminada exitosamente",
            "content": {
              "application/json": {
                "example": {
                  "message": "Sucursal eliminada exitosamente"
                }
              }
            }
          },
          "404": {
            "description": "Sucursal no encontrada con el ID proporcionado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/order": {
      "get": {
        "tags": [
          " ** Endpoints Orden **"
        ],
        "summary": "Obtiene una lista de todas las órdenes",
        "description": "Este endpoint devuelve una lista completa de todas las órdenes registradas en el sistema",
        "operationId": "obtenerTodos",
        "responses": {
          "200": {
            "description": "Lista de órdenes obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "204": {
            "description": "No hay órdenes disponibles (La lista está vacía)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelOrder"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelOrder"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          " ** Endpoints Orden **"
        ],
        "summary": "Crea una nueva orden",
        "description": "Este endpoint permite crear una nueva orden en el sistema",
        "operationId": "crearPedido",
        "requestBody": {
          "description": "Datos de la orden a crear",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              },
              "examples": {
                "Orden ejemplo": {
                  "description": "Orden ejemplo",
                  "value": {
                    "orderNumber": "ORD-XYZ123",
                    "issueDate": "2025-07-14T14:45:00",
                    "totalAmount": 12000.5,
                    "clientId": 2,
                    "cartId": 2
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Orden creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/order/byClient/{clientId}": {
      "get": {
        "tags": [
          " ** Endpoints Orden **"
        ],
        "summary": "Obtiene órdenes por ID de cliente",
        "description": "Este endpoint devuelve una lista de órdenes asociadas a un cliente específico",
        "operationId": "obtenerPorUserId",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Órdenes obtenidas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron órdenes para el cliente especificado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelOrder"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelOrder"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/carts/orders/{id}": {
      "get": {
        "tags": [
          "Cart"
        ],
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/order/delete/{orderId}": {
      "delete": {
        "tags": [
          " ** Endpoints Orden **"
        ],
        "summary": "Elimina una orden por su ID",
        "description": "Este endpoint permite eliminar una orden específica utilizando su ID",
        "operationId": "delete_3",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Orden eliminada exitosamente",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Orden no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "cartItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          }
        }
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fidelityPoints": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderNumber": {
            "type": "string"
          },
          "issueDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "cart": {
            "$ref": "#/components/schemas/Cart"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "sku": {
            "type": "string"
          }
        }
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderNumber": {
            "type": "string"
          },
          "issueDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "cartId": {
            "type": "integer",
            "format": "int64"
          },
          "clientId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EntityModelOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderNumber": {
            "type": "string"
          },
          "issueDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "cart": {
            "$ref": "#/components/schemas/Cart"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "Branch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "region": {
            "type": "string"
          }
        }
      },
...
      "CollectionModelEntityModelOrder": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "orderList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    }
  }
}